Update common file and git ignore.

+++ common.ps1
@@ -1,19 +1,37 @@
 #common functions for scripts
+
 function BuildTemplate ($tempfolder, $name, $hash, $url, $version, $description) {
     Write-Host "Building `"$name`" templates..."
     New-Item -ItemType Directory -Path "${tempfolder}/tools" -ErrorAction Ignore | Out-Null
 
-    $citemplate = Get-Content "${templates}/${name}_chocolateyInstall.ps1.template" -Raw
-
-    $citemplate = $citemplate -replace "%hash%", "$hash"
-    $citemplate = $citemplate -replace "%downloadurl%", "$url"
-
-    $citemplate | Out-File "${tempfolder}/tools/chocolateyInstall.ps1"
-
+    Write-Host "Building nuspec template..."
     $nstemplate = Get-Content "${templates}/${name}.nuspec.template" -Raw
     $nstemplate = $nstemplate -replace "%fileversion%", "$version"
     $nstemplate = $nstemplate -replace "%description%", "$description"
+    $nstemplate = $nstemplate -replace "%hash%", "$hash"
+    $nstemplate = $nstemplate -replace "%downloadurl%", "$url"
     $nstemplate | Out-File "$tempfolder/${name}.nuspec"
+
+    $files = Get-ChildItem -Path $templates -Filter "${name}_*"
+    foreach ($file in $files) {
+        if ($file.Name.EndsWith(".template")) {
+            $outfilename = ($file.Name -replace ".{9}$" -replace ".*_")
+            Write-Host "Building template '$outfilename'"
+            $templater = Get-Content "${templates}/$($file.Name)" -Raw
+
+            $templater = $templater -replace "%fileversion%", "$version"
+            $templater = $templater -replace "%description%", "$description"
+            $templater = $templater -replace "%hash%", "$hash"
+            $templater = $templater -replace "%downloadurl%", "$url"
+
+            
+            $templater | Out-File "${tempfolder}/tools/${outfilename}"
+        } else {
+            $newfilename = $file.Name  -replace ".*_"
+            Write-Host "Copying file '$newfilename'"
+            Copy-Item "${templates}/$($file.Name)" "${tempfolder}/tools/$newfilename"
+        }
+    }
 }
 
 function HashAndSizeFromFileURL ($url) {
@@ -31,12 +49,34 @@ function HashAndSizeFromFileURL ($url) {
 }
 
 function PackAndClean ($tempfolder) {
-    Write-Host "Pack, Push and Clean `"$tempfolder`""
+    if ("$env:debug" -ne "true") {
+        Write-Host "Pack, Push and Clean `"$tempfolder`""
+    } else {
+        Write-Host "Pack `"$tempfolder`""
+    }
     Set-Location -Path $tempfolder
     choco pack
-    choco push --api-key=$Env:CKEY
+    if ("$env:debug" -ne "true") {
+        choco push --api-key=$Env:CKEY
+    }
     Set-Location -Path $temp
-    Remove-Item -Path $tempfolder -Recurse
+    if ("$env:debug" -ne "true") {
+        Remove-Item -Path $tempfolder -Recurse
+    }
+}
+
+function JoinPath ($path) {
+    $joinedpathstring += $path[0]
+    for ($i=1;$i -lt $path.length; $i++) {
+        if ("" -ne ("{0}" -f $path[$i]).Trim()) {
+            if (!$path[$i].StartsWith("/")) {
+                $joinedpathstring += "/" + $path[$i]
+            } else {
+                $joinedpathstring += $path[$i]
+            }
+        }
+    }
+    $joinedpathstring 
 }
 
 function GetLastVersion ($verfile) {
@@ -45,6 +85,12 @@ function GetLastVersion ($verfile) {
 
 function GetFileSize ($bytesize) {
     "{0:N2}MB" -f ($bytesize / 1MB)
+}
+
+function ConvertDashVersion ($version, $dashpostfix) {
+    $splitver = $version.Split('-')
+    $res = $splitver[0] + "-" + $dashpostfix + $splitver[1]
+    $res
 }
 
 function NotePackageUpdateMsg ($version, $verfile, $message) {
